+++ src/collective/iconifiedcategory/browser/tabview.py
@@ -123,8 +123,8 @@ class CategoryColumn(column.GetAttrColumn):
     def renderCell(self, obj):
         category_title = safe_unicode(obj.category_title)
         if obj.subcategory_title:
-            category_title = "{0} / {1}".format(category_title,
-                                                safe_unicode(obj.subcategory_title))
+            category_title = u"{0} / {1}".format(category_title,
+                                                 safe_unicode(obj.subcategory_title))
         return category_title